Rename matchers module to literals
