+++ frontend/src/metabase-lib/v1/expressions/helper-text-strings.ts
@@ -9,7 +9,7 @@ import type {
 import type Database from "metabase-lib/v1/metadata/Database";
 import type { Expression } from "metabase-types/api";
 
-import { isLiteral } from "./matchers";
+import { isLiteral } from "./literal";
 
 const getDescriptionForNow: HelpTextConfig["description"] = (
   database,
