+++ frontend/src/metabase-lib/v1/expressions/index.ts
@@ -1,4 +1,3 @@
-export * from "./matchers";
 export * from "./config";
 export * from "./identifier";
 export * from "./formatter";
